# HashMap/HashSet optimization: fast modulo