This utilizes a technique researched by Daniel Lemire et al. (https://arxiv.org/abs/1902.01961) to calculate x mod d (where d is a constant) a lot faster than the regular expensive modulo operation %. Since the possible sizes of the hashtable are stored as constants in a lookup table we can precompute the required inverse elements and store them in a lookup table as well.
I came across this technique when I read about some optimizations of Boost's unordered_map for the upcoming 1.80 release.

The performance improvements are quite significant with up to **40%** (int - insertion), especially when the hash function isn't the bottleneck.

Benchmark with 200.000 elements (average of 200 runs)

Linux-Notebook: (Intel i5-8250U) built with GCC 12.1:
|int|insert|find|erase|  
|---|---|---|---|  
|HashMap| 37.789645ms |6.06988ms |17.559695ms |  
|HashMap (fastmod) |27.10462ms |4.26861ms |12.853815ms |  

|Vector2i|insert|find|erase|  
|---|---|---|---|  
|HashMap| 40.99277ms |7.48687ms |18.31907ms |
|HashMap (fastmod) |31.01118ms |5.6226ms |15.61076ms |

Desktop PC (AMD Ryzen7 3700X) built with MinGW 10.0(GCC 12.1):
|int key|insert|find|erase|
|---|---|---|---|
|HashMap| 26ms |5.2735ms |15.616ms |
|HashMap (fastmod) |19.3191ms |4.4281ms |14.8263ms |

|Vector2i key|insert|find|erase|
|---|---|---|---|
|HashMap| 26.224ms |7.1828ms |17.4173ms |
|HashMap (fastmod) |20.1043ms |6.5907ms |16.7279ms |

The results on the same machine with MSVC 14.3 are slightly worse, but still faster than before:
|int key|insert|find|erase|
|---|---|---|---|
|HashMap| 26.2993ms |6.8913ms |15.9505ms |
|HashMap (fastmod) |20.5731ms |6.9024ms |15.7328ms |

|Vector2i key|insert|find|erase|
|---|---|---|---|
|HashMap| 27.059ms |10.605ms |18.1936ms |
|HashMap (fastmod) |20.4513ms |9.6771ms |16.7354ms |

Detailed changes:
- Use fastmod for HashMap and HashSet (could also be applied to #61083)
    - Fallback to regular modulo if the required type/intrinsic isn't supported